Clean up built-in interfaces
